+++ x/ecocredit/server/basket/query_balance.go
@@ -6,8 +6,9 @@ import (
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 
+	"cosmossdk.io/errors"
+
 	"github.com/cosmos/cosmos-sdk/orm/types/ormerrors"
-	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 
 	baskettypes "github.com/regen-network/regen-ledger/x/ecocredit/basket"
 )
@@ -20,14 +21,14 @@ func (k Keeper) BasketBalance(ctx context.Context, request *baskettypes.QueryBas
 	basket, err := k.stateStore.BasketTable().GetByBasketDenom(ctx, request.BasketDenom)
 	if err != nil {
 		if ormerrors.IsNotFound(err) {
-			return nil, sdkerrors.Wrapf(err, "basket %s not found", request.BasketDenom)
+			return nil, errors.Wrapf(err, "basket %s not found", request.BasketDenom)
 		}
-		return nil, sdkerrors.Wrapf(err, "failed to get basket %s", request.BasketDenom)
+		return nil, errors.Wrapf(err, "failed to get basket %s", request.BasketDenom)
 	}
 
 	found, err := k.coreStore.BatchTable().HasByDenom(ctx, request.BatchDenom)
 	if err != nil {
-		return nil, sdkerrors.Wrapf(err, "failed to get credit batch %s", request.BatchDenom)
+		return nil, errors.Wrapf(err, "failed to get credit batch %s", request.BatchDenom)
 	}
 	if !found {
 		return nil, ormerrors.NotFound.Wrapf("credit batch %s not found", request.BatchDenom)
@@ -38,7 +39,7 @@ func (k Keeper) BasketBalance(ctx context.Context, request *baskettypes.QueryBas
 		if ormerrors.IsNotFound(err) {
 			return &baskettypes.QueryBasketBalanceResponse{Balance: "0"}, nil
 		}
-		return nil, sdkerrors.Wrapf(err, "failed to get basket balance for %s", request.BasketDenom)
+		return nil, errors.Wrapf(err, "failed to get basket balance for %s", request.BasketDenom)
 	}
 
 	return &baskettypes.QueryBasketBalanceResponse{Balance: balance.Balance}, nil